requirements.txt
langchain>=0.0.300
openai>=1.0.0
chromadb>=0.4.14
python-dotenv>=1.0.0
tiktoken>=0.5.1
streamlit
langchain-openai
pymysql
pandas